HALT inserted, various corrections

+++ src/ch/ntb/inf/deep/unsafe/US.java
@@ -53,7 +53,7 @@ public class US {
 	}
 
 	/** get content of floating point register */
-	public static int GETFPR(int reg) {
+	public static long GETFPR(int reg) {
 		return 0;
 	}
